% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents21}
\alias{accidents21}
\title{Car accidents in UK in 2021}
\format{
this data frame has 101.046 rows and the following 7 columns:
\describe{
\item{y}{the count for the 2-week period.}
\item{trt}{treatment, "placebo" or "progabide"}
\item{post}{post treatment. 0 for no, 1 for yes}
\item{subject}{subject id}
\item{tj}{time}
}
}
\source{
\url{https://www.data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data}
}
\usage{
accidents21
}
\description{
TODO: write a good description and document this properly
}
\keyword{datasets}
