% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{BARPLOT1}
\alias{BARPLOT1}
\alias{BARPLOT2}
\title{BARPLOT1}
\usage{
BARPLOT1(dataframe, var1, var2)

BARPLOT2(dataframe, var1, var2)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var1}{name of column in \code{dataframe}}

\item{var2}{name of column in \code{dataframe}}
}
\value{
\code{BARPLOT1()} returns a barplot

\code{BARPLOT2()} returns a barplot with facets.
}
\description{
\code{BARPLOT1()} generates a classical barplot

\code{BARPLOT2()} is similar to \code{BARPLOT1()} with an additional facet
given with \code{facet_wrap()}
}
\details{
With \code{var1} on the x-axis \code{BARPLOT1()} counts the total amount of the variable and
it is displayed on the y-axis. With \code{var2}, \code{BARPLOT1()} splits the bars in smaller colored bars,
that represent the amount of observations regarding the second variable.

In extension to \code{BARPLOT1()}, \code{BARPLOT2()} divides the plot in more
subplots to gain more visibility.
}
\examples{
BARPLOT1(euro_startups, status, country_code)
\dontrun{
# var1 and var2 must be passed as they are, without quotes or apostrophes.
BARPLOT1(euro_startups, "status", "country_code")
# throws an error.
}
BARPLOT2(euro_startups, status, country_code)
\dontrun{
 # var1 and var2 must be passed as they are, without quotes or apostrophes.
BARPLOT2(euro_startups, "status", "country_code")
# throws an error.
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_bar}}

\code{\link[katws]{BARPLOT1}} is a extension of this function that has additional facets.
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{facet_wrap}}
}
