% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{BARPLOT1}
\alias{BARPLOT1}
\title{BARPLOT1}
\usage{
BARPLOT1(dataframe, var1, var2)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var1}{name of column in \code{dataframe}}

\item{var2}{name of column in \code{dataframe}}
}
\value{
\code{BARPLOT1()} returns a barplot
}
\description{
\code{BARPLOT1()} generates a classical barplot
}
\details{
With \code{var1} on the x-axis \code{BARPLOT1()} counts the total amount of the variable and
it is displayed on the y-axis. With \code{var2}, \code{BARPLOT1()} splits the bars in smaller colored bars,
that represent the amount of observations regarding the second variable.
}
\examples{
BARPLOT1(euro_startups, status, country_code)
\dontrun{
# var1 and var2 must be passed as they are, without quotes or apostrophes.
BARPLOT1(euro_startups, "status", "country_code")
# throws an error.
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_bar}}
}
