% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{BARPLOT2}
\alias{BARPLOT2}
\title{BARPLOT2}
\usage{
BARPLOT2(dataframe, var1, var2)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var1}{name of column in \code{dataframe}}

\item{var2}{name of column in \code{dataframe}}
}
\value{
\code{BARPLOT2()} returns a barplot with facets.
}
\description{
\code{BARPLOT2()} is similar to \code{BARPLOT1()} with an additional facet
given with \code{facet_wrap()}
}
\details{
In extension to \code{BARPLOT1()}, \code{BARPLOT2()} divides the plot in more
subplots to gain more visibility.
}
\examples{
BARPLOT2(euro_startups, status, country_code)
\dontrun{
 # var1 and var2 must be passed as they are, without quotes or apostrophes.
BARPLOT2(euro_startups, "status", "country_code")
# throws an error.
}
}
\seealso{
\code{\link[katws]{BARPLOT1}} is a extension of this function that has additional facets.
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{facet_wrap}}
}
