% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_desc.R
\name{contingency_table}
\alias{contingency_table}
\title{Cross Tabulation}
\usage{
contingency_table(dataframe, var1, var2)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var1}{name of column in \code{dataframe}}

\item{var2}{name of column in \code{dataframe}}
}
\value{
\code{contingency_table()} returns a contingency table represented as
an object of type tibble
}
\description{
\code{contingency_table()} builds a contingency table of the counts at
each combination of values from two columns of a data frame.
}
\details{
the first column is given by \code{var1} and the following columns are
named by the unique values of \code{var2}. Each cell represents the count of
\code{var1}-\code{var2} pairs for the corresponding values.
}
\examples{
contingency_table(accidents21, accident_severity, urban_or_rural_area)
\dontrun{
# var1 and var2 must be passed as they are, without quotes or apostrophes.
contingency_table(accidents21, "accident_severity", "urban_or_rural_area")
# throws an error.
}
}
\seealso{
\code{\link[katws]{contingency_table_scale}} is a extension of this function
that gives the percentage the frequency of each pair among all pairs
}
