% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_factory.R
\name{disc}
\alias{disc}
\title{FUNCTION_TITLE}
\usage{
disc(lower_bound, upper_bound, nbins, value)
}
\arguments{
\item{lower_bound}{PARAM_DESCRIPTION}

\item{upper_bound}{PARAM_DESCRIPTION}

\item{nbins}{PARAM_DESCRIPTION}

\item{value}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
