% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_factory.R
\name{discretise}
\alias{discretise}
\title{FUNCTION_TITLE}
\usage{
discretise(mean, variance, nbins, vec)
}
\arguments{
\item{mean}{PARAM_DESCRIPTION}

\item{variance}{PARAM_DESCRIPTION}

\item{nbins}{PARAM_DESCRIPTION}

\item{vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
