% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euro_startups}
\alias{euro_startups}
\title{Startups in Western Europe}
\format{
this data frame has 7167 rows and the following 3 columns:
\describe{
\item{y}{the count for the 2-week period.}
\item{trt}{treatment, "placebo" or "progabide"}
\item{post}{post treatment. 0 for no, 1 for yes}
\item{subject}{subject id}
\item{tj}{time}
}
}
\source{
\url{https://www.kaggle.com/datasets/yanmaksi/big-startup-secsees-fail-dataset-from-crunchbase}
}
\usage{
euro_startups
}
\description{
TODO: write a good description and document this properly
}
\keyword{datasets}
