% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{HEATMAPS}
\alias{HEATMAPS}
\title{HEATMAPS}
\usage{
HEATMAPS(dataframe, var1, var2)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var1}{name of column in \code{dataframe}}

\item{var2}{name of column in \code{dataframe}}
}
\value{
\code{HEATMAPS()} returns a heatmap
}
\description{
\code{HEATMAPS()} is a function that generates a heatmap as a plot.
}
\details{
\code{HEATMAPS()} has a similar functionality like \code{COUNTPLOT()}, but
it differs only in visualization. It doesn't plot points,
but colored areas, whose colors visualize the amount of observations of given combinations.
}
\examples{
HEATMAPS(euro_startups, status, country_code)
\dontrun{
# var1 and var2 must be passed as they are, without quotes or apostrophes.
HEATMAPS(euro_startups, "status", "country_code")
# throws an error.
}
}
\seealso{
\code{\link[katws]{COUNTPLOT}} is similar to \code{COUNTPLOT()}, it differs only in visualization.
\code{\link[dplyr]{count}}, \code{\link[ggplot]{ggplot}}, \code{\link[ggplot]{geom_tile}}
}
