% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_desc.R
\name{how_many}
\alias{how_many}
\title{FUNCTION_TITLE}
\usage{
how_many(dataframe, variable)
}
\arguments{
\item{dataframe}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{count}}
}
