% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_factory.R
\name{jobs_colors_independent}
\alias{jobs_colors_independent}
\title{FUNCTION_TITLE}
\usage{
jobs_colors_independent(nrows = 1000, seed = 42)
}
\arguments{
\item{nrows}{PARAM_DESCRIPTION, Default: 1000}

\item{seed}{PARAM_DESCRIPTION, Default: 42}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
\code{\link[stats]{Uniform}}, \code{\link[stats]{Binomial}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{recode}}
}
