% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katws-package.R
\docType{package}
\name{katws-package}
\alias{katws}
\alias{katws-package}
\title{katws: katws}
\description{
Tools for numerical, graphical and statistical analysis of categorical variable.
}
\author{
\strong{Maintainer}: Ștefan Dan \email{stefan.dan@tum.de}

Authors:
\itemize{
  \item Teodor-Ștefan Ghidersa
}

}
\keyword{internal}
