% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_factory.R
\name{multivariateNormal}
\alias{multivariateNormal}
\title{FUNCTION_TITLE}
\usage{
multivariateNormal(nrows = 1000, mean, cov_matrix, seed = 42)
}
\arguments{
\item{nrows}{PARAM_DESCRIPTION, Default: 1000}

\item{mean}{PARAM_DESCRIPTION}

\item{cov_matrix}{PARAM_DESCRIPTION}

\item{seed}{PARAM_DESCRIPTION, Default: 42}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
\code{\link[MASS]{mvrnorm}}
}
