% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PIECHART}
\alias{PIECHART}
\title{PIECHART}
\usage{
PIECHART(dataframe, var)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var}{name of column in \code{dataframe}}
}
\value{
returns a piechart.
}
\description{
\code{PIECHART()} is a function that generates a piechart as a plot.
}
\details{
\code{PIECHART()} calculates the percentage of the observations of \code{var1}
and displays it onto the plot. The piechart is a barplot with a polar-transformation.
}
\examples{
PIECHART(euro_startups, status)
\dontrun{
PIECHART(euro_startups, "status")
# throws an error.
}
}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summmarise}},
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_bar}},
\code{\link[ggplot2]{coord_polar}},\code{\link[ggplot2]{geom_text}},
\code{\link[ggplot2]{labs}}
}
