% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_factory.R
\name{save_as}
\alias{save_as}
\title{FUNCTION_TITLE}
\usage{
save_as(dataframe, format, path)
}
\arguments{
\item{dataframe}{PARAM_DESCRIPTION}

\item{format}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readr]{write_delim}}
}
