% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_methods.R
\name{test_independence}
\alias{test_independence}
\title{FUNCTION_TITLE}
\usage{
test_independence(dataframe, var1, var2, sample_size, reps, visualize = TRUE)
}
\arguments{
\item{dataframe}{PARAM_DESCRIPTION}

\item{var1}{PARAM_DESCRIPTION}

\item{var2}{PARAM_DESCRIPTION}

\item{sample_size}{PARAM_DESCRIPTION}

\item{reps}{PARAM_DESCRIPTION}

\item{visualize}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{pull}}
\code{\link[infer]{rep_sample_n}}, \code{\link[infer]{specify}}, \code{\link[infer]{hypothesize}}, \code{\link[infer]{generate}}, \code{\link[infer]{calculate}}
}
