% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_desc.R
\name{contingency_table}
\alias{contingency_table}
\title{FUNCTION_TITLE}
\usage{
contingency_table(dataframe, var1, var2)
}
\arguments{
\item{dataframe}{PARAM_DESCRIPTION}

\item{var1}{PARAM_DESCRIPTION}

\item{var2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
\code{\link[tidyr]{spread}}
}
