% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_desc.R
\name{contingency_table_scale}
\alias{contingency_table_scale}
\title{Cross Tabulation with Percentages}
\usage{
contingency_table_scale(dataframe, var1, var2, all = FALSE)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var1}{name of column in \code{dataframe}}

\item{var2}{name of column in \code{dataframe}}

\item{all}{whether the percentages are relative to the entire \code{dataframe}
(\code{all = TRUE}) or to \code{var1}.}
}
\value{
\code{contingency_table_scale()} returns a contingency table represented as
an object of type tibble.
}
\description{
\code{contingency_table_scale()} builds a contingency table of the
percentages of each pair. If the parameter \code{all} is set to TRUE, then the
percentages are relative to the entire data frame. Otherwise, they are relative
to each value given in the \code{var1} parameter.
}
\details{
The first column is given by \code{var1} and the following columns are
named by the unique values of \code{var2}. Each cell represents the percentage of
\code{var1}-\code{var2} pairs for the corresponding values.
}
\examples{
contingency_table_scale(euro_startups, country_code, status, all = FALSE)
contingency_table_scale(euro_startups, country_code, status, all = TRUE)
\dontrun{
# var1 and var2 must be passed as they are, without quotes or apostrophes.
contingency_table_scale(euro_startups, "country_code", "status", all = FALSE)
# throws an error.
}
}
\seealso{
\code{\link[katws]{contingency_table}}
}
