% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{COUNTPLOT}
\alias{COUNTPLOT}
\title{COUNTPLOT}
\usage{
COUNTPLOT(dataframe, var1, var2, stat)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var1}{name of column in \code{dataframe}}

\item{var2}{name of column in \code{dataframe}}

\item{stat}{a char for \code{prop} or \code{count}}
}
\value{
\code{COUNTPLOT()} returns a countplot with either a \code{prop}-parameter
or \code{count}-parameter.
}
\description{
\code{COUNTPLOT()} generates a countplot as a plot.
}
\details{
\code{COUNTPLOT()} draws a point at each combination of values from the two
variables \code{var1} and \code{var2}. The size of the points is mapped to the number
of observations with this combinations of values. Meaning: rare observations
is equal to small points and frequent observations is equal to large points.
With \code{stat} you can decide whether to use \code{prop} or \code{count} as parameter, so
that you have two possibilities of plots.
}
\examples{
COUNTPLOT(euro_startups, status, country_code, "prop")
\dontrun{
# var1 and var2 must be passed as they are, without quotes or apostrophes.
# stat must be passed as either "prop", or "count".
COUNTPLOT(euro_startups, "status", "country_code", prop)
# throws an error.
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot]{geom_count}},
\code{\link[ggplot]{scale_size_area}}
}
