% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euro_startups}
\alias{euro_startups}
\title{Start-ups in Western Europe}
\format{
This tibble contains 7167 observations and the following 3 columns:
\describe{
\item{\code{category_list}}{Field(s) of activity of the start-up (saved as chr)}
\item{\code{status}}{current status of the start-up
(Factor, with levels "operating", "closed", "acquired")}
\item{\code{country_code}}{the country in which the start-up was founded
(Factor, with levels: \cr
"AUT" = Austria \cr
"BEL" = Belgium \cr
"CHE" = Switzerland \cr
"DEU" = Germany \cr
"FRA" = France \cr
"GBR" = Great Britain \cr
"IRL" = Ireland \cr
"SWE" = Sweden)}
}
}
\source{
The data originates from a bigger dataset,
\verb{Startup Success/Fail Dataset from Crunchbase}, which can be found on kaggle under
the following link:
\url{https://www.kaggle.com/datasets/yanmaksi/big-startup-secsees-fail-dataset-from-crunchbase}
}
\usage{
euro_startups
}
\description{
This tibble contains the areas of expertise and the current status
of start-ups founded in eight Western European countries.
}
\keyword{datasets}
