% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katws-package.R
\docType{package}
\name{katws-package}
\alias{katws}
\alias{katws-package}
\title{katws: The katws package does interesting stuff}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
