% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_methods.R
\name{test_independence}
\alias{test_independence}
\title{Independence Test}
\usage{
test_independence(dataframe, var1, var2, sample_size, reps, illustrate = TRUE)
}
\arguments{
\item{dataframe}{a data frame or data frame extension (e.g. a tibble)}

\item{var1}{name of column in \code{dataframe} that stores the first event}

\item{var2}{name of column in \code{dataframe} that stores the second event}

\item{sample_size}{the size of the sample used in the resampling method}

\item{reps}{number of repetitions in the resampling method}

\item{illustrate}{if TRUE, prints the result of the theoretical method and
plots the chi-squared value obtained theoretically
against the distribution from the resampling method and a reference chi-squared
distribution with the observed degrees of freedom, Default: TRUE}
}
\value{
A tibble containing a \code{stat} column of calculated chi-squared values
obtained in the resampling method.
}
\description{
The function tests whether two events given as columns of a
data frame are independent or not. It uses both a theoretical approach via
the chi-squared test from the \code{stats} package,
under the assumption that the values are independent,
and a resampling approach via functions from the \code{infer} package.
In the second method, we generate a sample of the events and permute the
observations of the second event against the ones of the first. For each
repetition, the chi-squared value is computed. The distribution of these values
approximates a chi-squared distribution.
Lastly, we compare the value obtained through the theoretical approach with the
distribution of the resampling method. If they are far from each other, we may
reject the Null-Hypothesis of independence.
}
\details{
The chi-squared test conducted in the theoretical approach might
return an inaccurate response if the relative proportions of the pairs of
observations are too skewed.
}
\examples{
stat1 <- test_independence(euro_startups, country_code, status, 1000, 1000, TRUE)
stat2 <- test_independence(accidents21, accident_severity, day_of_week, 2000, 1000, TRUE)
\dontrun{
# var1 and var2 must be passed as they are, without quotes or apostrophes.
 test_independence(euro_startups, "country_code", "status", 1000, 1000, TRUE)
# throws an error.
}
}
\seealso{
\code{\link[infer]{rep_sample_n}}, \code{\link[infer]{specify}},
\code{\link[infer]{hypothesize}}, \code{\link[infer]{generate}},
\code{\link[infer]{calculate}}
\code{\link[stats]{chisq.test}}
}
